Original change: https://android-review.googlesource.com/c/platform/system/update_engine/+/3248295

+++ payload_consumer/filesystem_verifier_action.h
@@ -109,7 +109,7 @@ class FilesystemVerifierAction : public InstallPlanAction {
 
   bool IsVABC(const InstallPlan::Partition& partition) const;
 
-  size_t GetPartitionSize() const;
+  uint64_t GetPartitionSize() const;
 
   // When the read is done, finalize the hash checking of the current partition
   // and continue checking the next one.
